check exception handling for mismatching parameter types
lua_typename seems to return nullptr or other valid strings when given
small negative values (-2, -3, ...) which represent index positions
instead of types. (although the original bug report said it segfaults -
which is possible too of course).

The std::string constructor throws if given a nullptr, but after rethrow
and nesting this ends as ExecutionErrorException. WrongTypeException
gets converted to a string for lua and then too ends up as
ExecutionErrorException - so we need to actually compare the exception
messages to make sure we got (the correct) WrongTypeException.